+++ CHANGELOG.md
@@ -33,6 +33,13 @@ Bugsnag Notifiers on other platforms.
      (Swift: `Bugsnag.clearMetadata(_ section)`)
      [#457](https://github.com/bugsnag/bugsnag-cocoa/pull/457)
 
+* Added `Bugsnag.getMetadata(_ section)`.  The behaviour is: calling with a valid section
+  name will return the metadata for that section if it exists, or `nil` if it does not exist.  Other,
+  similar functionality (e.g. `BugsnagConfiguration.getTab()` has been renamed and
+  had usage aligned with this change.
+  [#459](https://github.com/bugsnag/bugsnag-cocoa/pull/459)
+  
+
 ## Bug fixes
 
 * Fix possible report corruption when using `notify()` from multiple threads